Export and read in json format
